requirements.txt
streamlit>=1.32.0
plotly>=5.20.0
yfinance>=0.2.40
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
